+++ spring/src/main/java/org/sopt/spring/blog/controller/BlogController.java
@@ -1,0 +1,37 @@
+package org.sopt.spring.blog.controller;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.sopt.spring.common.success.SuccessMessage;
+import org.sopt.spring.common.success.SuccessStatusResponse;
+import org.sopt.spring.blog.service.BlogService;
+import org.sopt.spring.blog.dto.BlogCreateRequest;
+import org.sopt.spring.blog.dto.BlogTitleUpdateRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/v1")
+@RequiredArgsConstructor
+public class BlogController {
+    private final BlogService blogService;
+
+    @PostMapping("/blog")
+    public ResponseEntity<SuccessStatusResponse> createBlog(
+            @RequestHeader Long memberId,
+            @RequestBody BlogCreateRequest blogCreateRequest) {
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .header(
+                "Location", blogService.create(memberId, blogCreateRequest))
+                .body(SuccessStatusResponse.of(SuccessMessage.BLOG_CREATE_SUCCESS, null));
+    }
+    @PatchMapping("/blog/{blogId}/title")
+    public ResponseEntity updateBlogTitle(
+            @PathVariable Long blogId,
+            @Valid @RequestBody BlogTitleUpdateRequest blogTitleUpdateRequest
+    ){
+        blogService.updateTitle(blogId, blogTitleUpdateRequest);
+        return ResponseEntity.noContent().build();
+    }
+}
