+++ spring/src/main/java/org/sopt/spring/blog/dto/BlogCreateRequest.java
@@ -1,0 +1,7 @@
+package org.sopt.spring.blog.dto;
+
+public record BlogCreateRequest(
+        String title,
+        String description
+) {
+}
