+++ spring/src/main/java/org/sopt/spring/blog/repository/BlogRepository.java
@@ -1,0 +1,7 @@
+package org.sopt.spring.blog.repository;
+
+import org.sopt.spring.blog.domain.Blog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BlogRepository extends JpaRepository<Blog, Long> {
+}
