--- spring/src/main/java/org/sopt/spring/common/GlobalExceptionHandler.java
@@ -1,26 +1,0 @@
-package org.sopt.spring.common;
-
-import org.sopt.spring.common.dto.ErrorMessage;
-import org.sopt.spring.common.dto.ErrorResponse;
-import org.sopt.spring.exception.NotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import java.util.Objects;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(NotFoundException.class)
-    protected ResponseEntity<ErrorResponse> handleNotFoundException(NotFoundException e){
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(ErrorResponse.of(ErrorMessage.MEMBER_NOT_FOUND));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException e){
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ErrorResponse.of(HttpStatus.BAD_REQUEST.value(),
-                Objects.requireNonNull(e.getBindingResult().getFieldError().getDefaultMessage())));
-    }
-}
