+++ spring/src/main/java/org/sopt/spring/common/config/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package org.sopt.spring.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
