--- spring/src/main/java/org/sopt/spring/common/dto/ErrorMessage.java
@@ -1,16 +1,0 @@
-package org.sopt.spring.common.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorMessage {
-    MEMBER_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 사용자가 존재하지 않습니다"),
-    BLOG_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 블로그가 존재하지 않습니다")
-    ;
-    private final int status;
-    private final String message;
-}
-
