--- spring/src/main/java/org/sopt/spring/common/dto/SuccessStatusResponse.java
@@ -1,10 +1,0 @@
-package org.sopt.spring.common.dto;
-
-public record SuccessStatusResponse(
-        int status,
-        String message
-) {
-    public static SuccessStatusResponse of(SuccessMessage successMessage) {
-        return new SuccessStatusResponse(successMessage.getStatus(), successMessage.getMessage());
-    }
-}
