+++ spring/src/main/java/org/sopt/spring/common/exception/BusinessException.java
@@ -1,0 +1,10 @@
+package org.sopt.spring.common.exception;
+
+public class BusinessException extends RuntimeException {
+    private ErrorMessage errorMessage;
+
+    public BusinessException(ErrorMessage errorMessage) {
+        super(errorMessage.getMessage());
+        this.errorMessage = errorMessage;
+    }
+}
