+++ spring/src/main/java/org/sopt/spring/common/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package org.sopt.spring.common.exception;
+
+public record ErrorResponse(
+        int status,
+        String message
+) {
+    public static ErrorResponse of(int status, String message) {
+        return new ErrorResponse(status, message);
+    }
+
+    public static ErrorResponse of(ErrorMessage errorMessage) {
+        return new ErrorResponse(errorMessage.getStatus(), errorMessage.getMessage());
+    }
+}
