+++ spring/src/main/java/org/sopt/spring/common/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package org.sopt.spring.common.exception;
+
+public class NotFoundException extends BusinessException{
+    public NotFoundException(ErrorMessage errorMessage) {
+        super(errorMessage);
+    }
+}
