+++ spring/src/main/java/org/sopt/spring/common/exception/NotHaveBlog.java
@@ -1,0 +1,7 @@
+package org.sopt.spring.common.exception;
+
+public class NotHaveBlog extends BusinessException{
+    public NotHaveBlog(ErrorMessage errorMessage) {
+        super(errorMessage);
+    }
+}
