--- spring/src/main/java/org/sopt/spring/config/JpaAuditingConfig.java
@@ -1,10 +1,0 @@
-package org.sopt.spring.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-
-}
