--- spring/src/main/java/org/sopt/spring/controller/MemberController.java
@@ -1,51 +1,0 @@
-package org.sopt.spring.controller;
-
-
-import lombok.RequiredArgsConstructor;
-import org.sopt.spring.service.MemberService;
-import org.sopt.spring.service.dto.MemberCreateDto;
-import org.sopt.spring.service.dto.MemberFindDto;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/member")
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @PostMapping
-    public ResponseEntity createMember(
-            @RequestBody MemberCreateDto memberCreateDto
-    ) {
-        return ResponseEntity.created(URI.create(memberService.createMember(memberCreateDto)))
-                .build();
-    }
-
-    @GetMapping("/{memberId}")
-    public ResponseEntity<MemberFindDto> findMemberById(@PathVariable Long memberId) {
-        return ResponseEntity.ok(memberService.findMemberById(memberId));
-    }
-
-    @DeleteMapping("/{memberId}")
-    public ResponseEntity deleteMemberById(@PathVariable Long memberId){
-        memberService.deleteMemberById(memberId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/all")
-    public ResponseEntity<List<MemberFindDto>> findAllMembers() {
-        List<MemberFindDto> members = memberService.findAllMembers();
-        return ResponseEntity.ok(members);
-    }
-}
