--- spring/src/main/java/org/sopt/spring/domain/Part.java
@@ -1,10 +1,0 @@
-package org.sopt.spring.domain;
-
-public enum Part {
-    IOS,
-    SERVER,
-    ANDROID,
-    WEB,
-    PLAN,
-    DESIGN
-}
