--- spring/src/main/java/org/sopt/spring/domain/Post.java
@@ -1,21 +1,0 @@
-package org.sopt.spring.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class Post extends BaseTimeEntity{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Blog blog;
-}
