--- spring/src/main/java/org/sopt/spring/exception/BusinessException.java
@@ -1,12 +1,0 @@
-package org.sopt.spring.exception;
-
-import org.sopt.spring.common.dto.ErrorMessage;
-
-public class BusinessException extends RuntimeException {
-    private ErrorMessage errorMessage;
-
-    public BusinessException(ErrorMessage errorMessage) {
-        super(errorMessage.getMessage());
-        this.errorMessage = errorMessage;
-    }
-}
