--- spring/src/main/java/org/sopt/spring/exception/NotFoundException.java
@@ -1,9 +1,0 @@
-package org.sopt.spring.exception;
-
-import org.sopt.spring.common.dto.ErrorMessage;
-
-public class NotFoundException extends BusinessException{
-    public NotFoundException(ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
