+++ spring/src/main/java/org/sopt/spring/member/domain/Part.java
@@ -1,0 +1,10 @@
+package org.sopt.spring.member.domain;
+
+public enum Part {
+    IOS,
+    SERVER,
+    ANDROID,
+    WEB,
+    PLAN,
+    DESIGN
+}
