+++ spring/src/main/java/org/sopt/spring/member/dto/MemberCreateDto.java
@@ -1,0 +1,10 @@
+package org.sopt.spring.member.dto;
+
+import org.sopt.spring.member.domain.Part;
+
+public record MemberCreateDto(
+        String name,
+        Part part,
+        int age
+) {
+}
