+++ spring/src/main/java/org/sopt/spring/member/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package org.sopt.spring.member.repository;
+
+import org.sopt.spring.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+}
