+++ spring/src/main/java/org/sopt/spring/post/repository/PostRepository.java
@@ -1,0 +1,7 @@
+package org.sopt.spring.post.repository;
+
+import org.sopt.spring.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
