--- spring/src/main/java/org/sopt/spring/repository/BlogRepository.java
@@ -1,7 +1,0 @@
-package org.sopt.spring.repository;
-
-import org.sopt.spring.domain.Blog;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BlogRepository extends JpaRepository<Blog, Long> {
-}
