--- spring/src/main/java/org/sopt/spring/repository/MemberRepository.java
@@ -1,8 +1,0 @@
-package org.sopt.spring.repository;
-
-import org.sopt.spring.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
