--- spring/src/main/java/org/sopt/spring/service/MemberService.java
@@ -1,57 +1,0 @@
-package org.sopt.spring.service;
-
-import jakarta.persistence.EntityNotFoundException;
-import lombok.RequiredArgsConstructor;
-import org.sopt.spring.common.dto.ErrorMessage;
-import org.sopt.spring.domain.Member;
-import org.sopt.spring.exception.NotFoundException;
-import org.sopt.spring.repository.MemberRepository;
-import org.sopt.spring.service.dto.MemberCreateDto;
-import org.sopt.spring.service.dto.MemberFindDto;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class MemberService {
-
-    private final MemberRepository memberRepository;
-
-    @Transactional
-    public String createMember(
-            MemberCreateDto memberCreateDto
-    ) {
-        Member member = Member.create(memberCreateDto.name(), memberCreateDto.part(), memberCreateDto.age());
-        memberRepository.save(member);
-        return member.getId().toString();
-    }
-
-    public MemberFindDto findMemberById(Long memberId) {
-        return MemberFindDto.of(memberRepository.findById(memberId).orElseThrow(
-                () -> new EntityNotFoundException("ID에 해당하는 사용자가 존재하지 않습니다.")
-        ));
-    }
-
-    @Transactional
-    public void deleteMemberById(Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new EntityNotFoundException("ID에 해당하는 사용자가 존재하지 않습니다."));
-        memberRepository.delete(member);
-    }
-
-    public List<MemberFindDto> findAllMembers() {
-        List<Member> members = memberRepository.findAll();
-        return members.stream()
-                .map(MemberFindDto::of)
-                .collect(Collectors.toList());
-    }
-
-    public Member findById(Long memberId) {
-        return memberRepository.findById(memberId).orElseThrow(
-                () -> new NotFoundException(ErrorMessage.MEMBER_NOT_FOUND)
-        );
-    }
-}
