--- spring/src/main/java/org/sopt/spring/service/dto/BlogCreateRequest.java
@@ -1,7 +1,0 @@
-package org.sopt.spring.service.dto;
-
-public record BlogCreateRequest(
-        String title,
-        String description
-) {
-}
