--- spring/src/main/java/org/sopt/spring/service/dto/MemberCreateDto.java
@@ -1,10 +1,0 @@
-package org.sopt.spring.service.dto;
-
-import org.sopt.spring.domain.Part;
-
-public record MemberCreateDto(
-        String name,
-        Part part,
-        int age
-) {
-}
